pyproject.toml
[project]
name = "finetune"
version = "0.1.0"
description = "Fine-tune instruct models using Apple Silicon MLX"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mlx-lm==0.30.7",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "mlflow>=2.0",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "openai>=2.24.0",
    "python-dotenv>=1.2.1",
]
